+++ Assets/Scripts/Objects/People/Person.cs
@@ -7,12 +7,12 @@ public abstract class Person : MonoBehaviour
     [SerializeField] protected float damage;
     [SerializeField] protected float attackRange;
     [SerializeField] protected float attackSpeed;
-    [SerializeField] protected int aggroVal;
+    [SerializeField] protected float aggroVal;
     [SerializeField] protected float damageArea;
     [SerializeField] protected bool isFriendly;
     [SerializeField] protected bool[] powerUpList = new bool[5]; // [strength, speed, shield, -, -]
 
-    public int AggroVal => aggroVal;
+    public float AggroVal => aggroVal;
 
     public bool IsFriendly => isFriendly;
